More refactoring of code
