+++ src/scenes/idle-game-scene/IdleGameScene.ts
@@ -8,19 +8,20 @@ import './game-objects';
 import CampManager from './home-managers/CampManager';
 import HomeManager from './home-managers/HomeManager';
 import MarketManager from './market-manager/MarketManager';
+import Log from './resources/Log';
 import { ResourceType } from './resources/Resource';
 import CoalManager from './resources/resource-managers/CoalManager';
 import LogManager from "./resources/resource-managers/LogManager";
-import CoalUpgradeManager from './upgrade-managers/resource-upgrade-managers/CoalUpgradeManager';
-import LogUpgradeManager from './upgrade-managers/resource-upgrade-managers/LogUpgradeManager';
+import CoalUpgradeManager from './resources/upgrades/upgrade-managers/CoalUpgradeManager';
+import LogUpgradeManager from './resources/upgrades/upgrade-managers/LogUpgradeManager';
 
 export class IdleGameScene extends Phaser.Scene {
 	// The home available to the user
 	public homeManager: HomeManager;
 	
 	// All the resources available to the user
-	public logManager = new LogManager();
-	public coalManager = new CoalManager();
+	// public logManager = new LogManager();
+	// public coalManager = new CoalManager();
 	
 	public marketManager: MarketManager;
 
@@ -33,32 +34,37 @@ export class IdleGameScene extends Phaser.Scene {
 	}
 
 	public create() {
-		const gameWidth = getGameWidth(this);
-		const gameHeight = getGameHeight(this);
-
-		// TODO: Create a custom GameObject to display this
-		// TODO: All this should also be inside the MarketManager class and will be create on new MarketManager();
-		// Market Area
-		this.marketManager = new MarketManager();
-		this.data.set('marketManager', this.marketManager as MarketManager);
-		this.add.market(this.marketManager, (gameWidth / 2) + 10, 10, ((gameWidth / 2) - 20), gameHeight - 20);
-
-		// Set all the upgrade managers this market can handle
-		this.marketManager.setUpgradeManager(ResourceType.LOG, new LogUpgradeManager(this.logManager));
-		this.marketManager.setUpgradeManager(ResourceType.COAL, new CoalUpgradeManager(this.coalManager));
-
-		// HOME
-		this.homeManager = new CampManager(this.logManager);
+		// Testing
+		let logManager = new LogManager();
 		
-		this.add.home(this.homeManager, 200, 200);
 
-		this.add.logResourceCollector(this.logManager, 10, 10);
-		this.add.coalResourceCollector(this.coalManager, 200, 10);
+		// Done testing
+		// const gameWidth = getGameWidth(this);
+		// const gameHeight = getGameHeight(this);
+
+		// // TODO: Create a custom GameObject to display this
+		// // TODO: All this should also be inside the MarketManager class and will be create on new MarketManager();
+		// // Market Area
+		// this.marketManager = new MarketManager();
+		// this.data.set('marketManager', this.marketManager as MarketManager);
+		// this.add.market(this.marketManager, (gameWidth / 2) + 10, 10, ((gameWidth / 2) - 20), gameHeight - 20);
+
+		// // Set all the upgrade managers this market can handle
+		// this.marketManager.setUpgradeManager(ResourceType.LOG, new LogUpgradeManager(this.logManager));
+		// this.marketManager.setUpgradeManager(ResourceType.COAL, new CoalUpgradeManager(this.coalManager));
+
+		// // HOME
+		// this.homeManager = new CampManager(this.logManager);
+		
+		// this.add.home(this.homeManager, 200, 200);
+
+		// this.add.logResourceCollector(this.logManager, 10, 10);
+		// this.add.coalResourceCollector(this.coalManager, 200, 10);
 	}
 	
 	public update(time: number, delta: number) {
-		this.logManager.update(delta);
-		this.coalManager.update(delta);
-		this.homeManager.update(delta);
+		// this.logManager.update(delta);
+		// this.coalManager.update(delta);
+		// this.homeManager.update(delta);
 	}
 }
