+++ src/scenes/idle-game-scene/resources/Log.ts
@@ -1,4 +1,5 @@
 import Resource from "./Resource";
+import LogManager from "./resource-managers/LogManager";
 
 export default class Log extends Resource {
 	public constructor() {
