+++ src/scenes/idle-game-scene/resources/Resource.ts
@@ -1,3 +1,5 @@
+import ResourceManager from "./resource-managers/ResourceManager";
+
 export enum ResourceType {
 	LOG,
 	COAL,
